// apps/backend/src/model/kanel/public/SEOServiceInstance.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ServiceInstanceId } from './ServiceInstance';

/** Identifier type for public.SEO_ServiceInstance */
export type SEOServiceInstanceLanguage = string & { __brand: 'public.SEO_ServiceInstance' };

/** Represents the table public.SEO_ServiceInstance */
export default interface SEOServiceInstance {
  service_instance_id: ServiceInstanceId;

  language: SEOServiceInstanceLanguage;

  meta_title: string;

  meta_description: string;
}

/** Represents the initializer for the table public.SEO_ServiceInstance */
export interface SEOServiceInstanceInitializer {
  service_instance_id: ServiceInstanceId;

  language: SEOServiceInstanceLanguage;

  meta_title: string;

  meta_description: string;
}

/** Represents the mutator for the table public.SEO_ServiceInstance */
export interface SEOServiceInstanceMutator {
  service_instance_id?: ServiceInstanceId;

  language?: SEOServiceInstanceLanguage;

  meta_title?: string;

  meta_description?: string;
}